feat(farm-card): Move bsc link into expandable section
