+++ src/components/ExpandableSectionButton/ExpandableSectionButton.tsx
@@ -11,6 +11,7 @@ const Wrapper = styled.div`
   display: flex;
   align-items: center;
   justify-content: center;
+  cursor: pointer;
 
   svg {
     fill: ${({ theme }) => theme.colors.primary};
